+++ src/main/java/kitchenpos/menu/domain/JpaMenuDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuDao extends JpaRepository<Menu, Long>, MenuDao {
