+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,9 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.menu.domain;
 
 import java.util.List;
-import kitchenpos.menu.repository.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.menu.domain.Menu;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
