+++ src/main/java/kitchenpos/product/domain/JpaProductDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.product.domain;
 
-import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductDao extends JpaRepository<Product, Long>, ProductDao {
