+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductResponse {
