+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -13,18 +13,18 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.menu.repository.MenuDao;
+import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
-import kitchenpos.dao.ProductDao;
+import kitchenpos.product.domain.ProductDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.dto.OrderChangeOrderStatusRequest;
 import kitchenpos.dto.OrderLineItemSaveRequest;
 import kitchenpos.dto.OrderResponse;
