+++ src/test/java/kitchenpos/menu/domain/MenuTest.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.common.UnitTest;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
